@@ -1,8 +1,9 @@
 package com.example.weatherkz.pojo;
 
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.Index;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
+import androidx.room.Entity;
+import androidx.room.Index;
 
 @Entity(primaryKeys = {"cityName", "temperature"}, indices = {@Index(value = {"cityName", "temperature"}, unique = true)})
 public class Weather {
