@@ -1,10 +1,10 @@
 package com.example.weatherkz.db;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.RoomDatabase;
-
 import com.example.weatherkz.pojo.QueryAndTime;
 import com.example.weatherkz.pojo.Weather;
+
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
 
 @Database(entities = {QueryAndTime.class, Weather.class}, version = 1)
 public abstract class WeatherDatabase extends RoomDatabase {
