@@ -1,13 +1,12 @@
 package com.example.weatherkz.di;
 
-import android.arch.lifecycle.ViewModel;
-
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import androidx.lifecycle.ViewModel;
 import dagger.MapKey;
 
 @Documented
