@@ -1,11 +1,11 @@
 package com.example.weatherkz.di;
 
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
 
 import com.example.weatherkz.repository.ViewModelFactory;
 import com.example.weatherkz.ui.MainViewModel;
 
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
 import dagger.Binds;
 import dagger.Module;
 import dagger.multibindings.IntoMap;
