@@ -2,10 +2,6 @@ package com.example.weatherkz.ui;
 
 import android.annotation.SuppressLint;
 import android.os.AsyncTask;
-import android.support.annotation.MainThread;
-import android.support.annotation.NonNull;
-import android.support.v7.util.DiffUtil;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -16,11 +12,16 @@ import com.example.weatherkz.pojo.Weather;
 
 import java.util.List;
 
+import androidx.annotation.MainThread;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.DiffUtil;
+import androidx.recyclerview.widget.RecyclerView;
+
 public class WeatherAdapter extends RecyclerView.Adapter<WeatherAdapter.WeatherVH> {
     private List<Weather> items;
     private int dataVersion = 0;
 
-    public WeatherAdapter() {
+    WeatherAdapter() {
     }
 
     @NonNull
@@ -47,7 +48,7 @@ public class WeatherAdapter extends RecyclerView.Adapter<WeatherAdapter.WeatherV
 
     @SuppressLint("StaticFieldLeak")
     @MainThread
-    public void replace(final List<Weather> update) {
+    void replace(final List<Weather> update) {
         dataVersion++;
         if (items == null) {
             if (update == null) {
